Move package nlp to model.gen.nlp
